add collaborations

+++ src/api/playlists/handler.js
@@ -19,7 +19,7 @@ module.exports = class PlaylistsHandler {
 
   async getPlaylistsHandler(req, h) {
     const playlists = await this._playlistsService.getPlaylists(req.auth.credentials.id);
-    console.log(playlists);
+
     return this._renderResponse(h, {
       data: {playlists},
     });
@@ -36,7 +36,7 @@ module.exports = class PlaylistsHandler {
 
   async postSongToPlaylistsHandler(req, h) {
     this._validator.validatePostSongOnPlaylistPayload(req.payload);
-
+    await this._playlistsService.verifyPlaylistAccess(req.params.id, req.auth.credentials.id);
     await this._playlistsService.verifyPlaylistOwner(req.params.id, req.auth.credentials.id);
     await this._songsService.verifySongIsExists(req.payload.songId);
     await this._playlistsService.addSongToPlaylists(req.params.id, req.payload);
@@ -49,6 +49,7 @@ module.exports = class PlaylistsHandler {
 
   async getSongsOnPlaylistsHandler(req, h) {
     await this._playlistsService.verifyPlaylistOwner(req.params.id, req.auth.credentials.id);
+    await this._playlistsService.verifyPlaylistAccess(req.params.id, req.auth.credentials.id);
     const playlist = await this._playlistsService.getSongsOnPlaylist(req.params.id, req.auth.credentials.id);
 
     return this._renderResponse(h, {
@@ -58,7 +59,7 @@ module.exports = class PlaylistsHandler {
 
   async deleteSongOnPlaylistHandler(req, h) {
     this._validator.validateDeleteSongOnPlaylistPayload(req.payload);
-
+    await this._playlistsService.verifyPlaylistAccess(req.params.id, req.auth.credentials.id);
     await this._playlistsService.verifyPlaylistOwner(req.params.id, req.auth.credentials.id);
     await this._songsService.verifySongIsExists(req.payload.songId);
     await this._playlistsService.deleteSongOnPlaylist(req.payload.songId);
